consolidated `joinClassNames` in `dom-node-class` subdirectory
