+++ public/leaflet-fullscreen/fullscreen-map.js
@@ -5,7 +5,7 @@ import {
 } from '../leaflet/leaflet-src.esm.js'
 
 import { controlAddedListener } from './control/control-added-listener.js'
-import { joinClassNames } from './join-class-names.js'
+import { joinClassNames } from './dom-node-class/join-class-names.js'
 import { mapLifecycleListener } from './map/map-lifecycle-listener.js'
 import { useBoolean } from './state/use-boolean.js'
 import { useLink } from './state/use-link.js'
